+++ front/components/skills/import/DetectedSkillsList.tsx
